requirements.txt
streamlit
langchain-groq
langchain-community
pypdf
sentence-transformers
langchain
langchain-openai
faiss-cpu
load_dotenv
